feat: updated documentation, added some needed files, worked on MaxHeap.cpp
