+++ CMakeLists.txt
@@ -1,7 +1,26 @@
 cmake_minimum_required(VERSION 3.28)
 project(SP24_SE_CSCI_C455_Xavbeat03_Homework_Assignment_4)
 
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
-add_executable(SP24_SE_CSCI_C455_Xavbeat03_Homework_Assignment_4
-        src/MaxHeap.cpp)
+include(FetchContent)
+FetchContent_Declare(
+        googletest
+        URL https://github.com/google/googletest/archive/03597a01ee50ed33e9dfd640b249b4be3799d395.zip
+)
+# For Windows: Prevent overriding the parent project's compiler/linker settings
+set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
+FetchContent_MakeAvailable(googletest)
+
+enable_testing()
+
+add_executable(
+        src/MaxHeap/MaxHeap.cpp
+        src/PriorityQueue/PriorityQueue.h
+)
+
+target_link_libraries(
+        gtest_discover_tests
+)
+
