+++ src/PriorityQueue/PriorityQueue.h
@@ -1,0 +1,14 @@
+//
+// Created by xavier on 3/12/24.
+//
+
+#ifndef SP24_SE_CSCI_C455_XAVBEAT03_HOMEWORK_ASSIGNMENT_4_PRIORITYQUEUE_H
+#define SP24_SE_CSCI_C455_XAVBEAT03_HOMEWORK_ASSIGNMENT_4_PRIORITYQUEUE_H
+
+
+class PriorityQueue {
+
+};
+
+
+#endif //SP24_SE_CSCI_C455_XAVBEAT03_HOMEWORK_ASSIGNMENT_4_PRIORITYQUEUE_H
